+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+1. Add `contains_children` method in `Focusable` trait to allow shadowing children in hit testing.
+
 ## [0.15.2] - 2026-07-08
 
 ### Changed